refactor deployment mutators
